apps/backend/requirements.txt
# --- Web Framework ---
fastapi==0.115.2
uvicorn==0.30.6

# --- Supabase Client ---
supabase==2.4.0
postgrest==0.16.4

# --- AI & NLP ---
openai==1.51.0
tiktoken==0.7.0

# --- Document Processing ---
PyPDF2==3.0.1
python-docx==1.1.0

# --- Utilities ---
requests==2.32.3
numpy==1.26.4
python-dotenv==1.0.1